Passing 24 out of 75 tests for PS6 GradingTests

+++ Spreadsheet/SpeadSheetTests/UnitTest1.cs
@@ -15,7 +15,7 @@ namespace SpeadSheetTests
 
             s.SetContentsOfCell("A1", "100");
             s.SetContentsOfCell("A2", "=A1+50");
-            Assert.AreEqual(s.GetCellValue("A2"), 150);
+            Assert.AreEqual(s.GetCellValue("A2"), (double)150);
 
             
         }
